+++ jsonargparse/__init__.py
@@ -67,4 +67,4 @@ __all__ += util.__all__
 __all__ += deprecated.__all__
 
 
-__version__ = '4.20.1'
+__version__ = '4.21.0'