+++ CHANGELOG.rst
@@ -19,6 +19,9 @@ Changed
 ^^^^^^^
 - Switched from ``setup.cfg`` to ``pyproject.toml`` for configuration.
 - Removed ``build_sphinx`` from ``setup.py`` and documented how to build.
+- Include enum members in error when invalid value is given
+  `pytorch-lightning#17247
+  <https://github.com/Lightning-AI/lightning/issues/17247>`__.
 
 
 v4.20.1 (2023-03-30)