+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
-v4.21.0 (2023-04-??)
+v4.21.0 (2023-04-21)
 --------------------
 
 Added